green_tech_backend/requirements.txt
# Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
setuptools>=65.5.0
wheel>=0.41.0

# Database (SQLite for development, PostgreSQL for production)
# psycopg2-binary==2.9.9  # Uncomment for PostgreSQL

# Authentication
djangorestframework-simplejwt==5.3.1
dj-rest-auth==5.0.1
django-model-utils==4.3.1

# Async Tasks
celery==5.3.6
redis==5.0.1

# File Storage
django-storages==1.14.2
boto3==1.34.13

# API Documentation
drf-yasg==1.21.7

# Code Quality
black==23.11.0
flake8==7.0.0
isort==5.13.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Localization
python-i18n==0.3.9
